Adding the Media Container's first draft.
I think the smarter thing to do would be to make it so when you click a card, a modal pops open with a larger image, the description. If it is a video, then perhaps the embedded video. This way you can dismiss the modal and remain on the site without having to navigate back.